+++ webapp-frontend/src/app/overall-control/overall-control.component.spec.ts
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,25 +17,28 @@
  * limitations under the License.
  * ========================LICENSE_END===================================
  */
- .ranconnect__section {
-}
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-.ranconnect__header {
-    text-align: center;
-    color: #432c85;
-    font-size: 50px;
-    font-weight: 200;
-    letter-spacing: .1em;
-    transform: translate(149 56);
-}
+import { OverallControlComponent } from './overall-control.component';
 
-.ranconnect-table {
-    width: 100%;
-    min-height: 100px;
-    margin-top: 10px;
-    background-color:transparent;
-}
+describe('OverallControlComponent', () => {
+  let component: OverallControlComponent;
+  let fixture: ComponentFixture<OverallControlComponent>;
 
-.disconnect-all-button {
-    float: right;
-}
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ OverallControlComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(OverallControlComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});