+++ docs/release-notes.rst
@@ -40,6 +40,7 @@ Version 1.0.4, 21 June 2019
 * Move mock admin screen user data to backend
 * Update App manager client to spec version 0.1.5
 * Rework admin table
+* Remove the RAN connection invocation link from left menu and move it to control screen
 
 Version 1.0.3, 28 May 2019
 --------------------------